+++ CHANGELOG.md
@@ -1,2 +1,7 @@
+# Akeeba Panopticon Connector for Joomla! 3.x version 1.0.1
+
+* ➕ Alternative token derivation method, working around plugins overriding database connection info in the frontend
+
 # Akeeba Panopticon Connector for Joomla! 3.x version 1.0.0
 
+* Initial release